requirements.txt
flask==3.1.1
line-bot-sdk==3.17.1
python-dotenv==1.1.0
schedule==1.2.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.172.0
gunicorn==21.2.0
supabase==2.16.0 